Refactor use of timeout for better test interaction and convert from Enzyme to RTL test
